requirements.txt
Django
leancloud-sdk>=1.0.9
pyopenssl
ndg-httpsclient
pyasn1